+++ migrations/up/014_create_cf_limit_value.sql
@@ -8,7 +8,7 @@ CREATE TABLE public.cf_limit_value (
 	id uuid NOT NULL DEFAULT uuid_generate_v4(),
 	law varchar(200) NOT NULL,
 	law_date timestamp NOT NULL,
-	value numeric(10, 4) NOT NULL,
+	value float8 NOT NULL,
 	created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	updated_at timestamp NULL,
 	e_tag uuid NOT NULL DEFAULT uuid_generate_v4(),
